fixed sh<defunct> from redis_manager's os.popen
