+++ config/slips.conf
@@ -388,7 +388,12 @@ use_p2p = no
 [Profiling]
 
 # cpu profiling
-cpu_profiler_enable = yes
+
+# enable cpu profiler [yes,no]
+cpu_profiler_enable = no
+# set profiler mode to [dev,live] -> dev for deterministic profiling, live for sampling data stream accessible from web interface
 cpu_profiler_mode = dev
+# set maximum output lines (dev mode only)
 cpu_profiler_output_limit = 20
+# set the wait time between sampling sequences in seconds (live mode only)
 cpu_profiler_sampling_interval = 20
