+++ slips.py
@@ -19,7 +19,6 @@
 
 import contextlib
 from slips_files.common.slips_utils import utils
-from slips_files.core.database.database import __database__
 from slips_files.common.config_parser import ConfigParser
 from slips_files.common.cpu_profiler import CPUProfiler
 from slips_files.common.imports import *
@@ -688,10 +687,14 @@ if __name__ == '__main__':
     slips = Main()
     cpuProfilerEnabled = slips.conf.get_cpu_profiler_enable() == 'yes'
     if cpuProfilerEnabled:
-        cpuProfiler = CPUProfiler(mode=slips.conf.get_cpu_profiler_mode(),
-                                  limit=slips.conf.get_cpu_profiler_output_limit(),
-                                  interval=slips.conf.get_cpu_profiler_sampling_interval())
-        cpuProfiler.start()
+        try:
+            cpuProfiler = CPUProfiler(mode=slips.conf.get_cpu_profiler_mode(),
+                                    limit=slips.conf.get_cpu_profiler_output_limit(),
+                                    interval=slips.conf.get_cpu_profiler_sampling_interval())
+            cpuProfiler.start()
+        except Exception as e:
+            print(e)
+            cpuProfilerEnabled = False
     if slips.args.stopdaemon:
         # -S is provided
         daemon = Daemon(slips)
