+++ slips_files/common/abstracts.py
@@ -133,3 +133,20 @@ class Module(ABC):
                 self.print(f'Problem in main() line {exception_line}', 0, 1)
                 self.print(traceback.format_exc(), 0, 1)
                 return True
+
+class ProfilerInterface(ABC):
+    @abstractmethod
+    def _create_profiler(self):
+        pass
+
+    @abstractmethod
+    def start(self):
+        pass
+
+    @abstractmethod
+    def stop(self):
+        pass
+
+    @abstractmethod
+    def print(self):
+        pass
