+++ slips_files/common/cpu_profiler.py
@@ -4,31 +4,13 @@ import threading
 import yappi
 import pstats
 
-from abc import ABC, abstractmethod
-
-#interface here for now, will move to separate file once I start memory profiler
-class ProfilerInterface(ABC):
-    @abstractmethod
-    def _create_profiler(self):
-        pass
-
-    @abstractmethod
-    def start(self):
-        pass
-
-    @abstractmethod
-    def stop(self):
-        pass
-
-    @abstractmethod
-    def print(self):
-        pass
+from slips_files.common.abstracts import ProfilerInterface
 
 class CPUProfiler(ProfilerInterface):
     def __init__(self, mode="dev", limit=20, interval=20):
         valid_modes = ["dev", "live"]
         if mode not in valid_modes:
-            raise Exception(mode + " is invalid, must be one of " + valid_modes)
+            raise Exception("cpu_profiler_mode = " + mode + " is invalid, must be one of " + str(valid_modes) + ", CPU Profiling will be disabled")
         if mode == "dev":
             self.profiler = DevProfiler(limit)
         if mode == "live":
@@ -38,10 +20,12 @@ class CPUProfiler(ProfilerInterface):
         self.profiler._create_profiler()
 
     def start(self):
+        print("CPU Profiler Started")
         self.profiler.start()
 
     def stop(self):
         self.profiler.stop()
+        print("CPU Profiler Ended")
 
     def print(self):
         self.profiler.print()
@@ -61,7 +45,7 @@ class DevProfiler(ProfilerInterface):
         self.profiler.stop()
 
     def print(self):
-        self.profiler.save()
+        self.profiler.save('output/result.json')
 
 class LiveProfiler(ProfilerInterface):
     def __init__(self, limit=20, interval=20):
